backend/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asyncpg==0.30.0
boto3==1.36.11
botocore==1.36.11
click==8.1.8
fastapi==0.115.7
greenlet==3.1.1
h11==0.14.0
idna==3.10
jmespath==1.0.1
pydantic==2.10.6
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
s3transfer==0.11.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.37
starlette==0.45.3
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
